+++ XServer/src/au/com/darkside/XServer/GContext.java
@@ -217,12 +217,12 @@ public class GContext extends Resource {
 	@Override
 	public void
 	processRequest (
-		Client			client,
-		byte			opcode,
-		int				arg,
-		int				bytesRemaining
+		Client		client,
+		byte		opcode,
+		byte		arg,
+		int			bytesRemaining
 	) throws IOException {
-		InputOutput		io = client.getInputOutput ();
+		InputOutput	io = client.getInputOutput ();
 
 		switch (opcode) {
 			case RequestCode.QueryFont:
